requirements.txt
requests
fastapi
strands-agents
strands-agents-tools
strands-agents-tools[rss]
aws-opentelemetry-distro
